Split the ThreadPool experiment into a header file and main source
